requirements.txt
autopep8==1.5.5
psycopg2==2.8.6
pycodestyle==2.6.0
python-dotenv==0.15.0
toml==0.10.2
XlsxWriter==1.3.7
